feat: write recieveAttack() and ship placing related methods
Write 'recieveAttaack(cordinate)' method which marks a block on the
gameboard as attacked, if a ship is present on the cordinate, deal
damage to it.
'areCordinatesValid(board, cordinates)' method checks if given
cordinates are valid or not.
'areCordinatesFree(cordinates)' method checks if any ship is already
present or given cordnates are overlapping any ship on the board.